# backend/openapi_schema.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "NEERON API",
    "description": "Precision Aquaculture Intelligence Platform",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/telemetry/latest": {
      "get": {
        "tags": [
          "Telemetry"
        ],
        "summary": "Get Latest Telemetry",
        "description": "Retrieves the latest pivoted environment metrics snapshot for a specific tank.",
        "operationId": "get_latest_telemetry_api_v1_telemetry_latest_get",
        "parameters": [
          {
            "name": "tank_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Tank Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponse_TelemetryLatestResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/telemetry/history": {
      "get": {
        "tags": [
          "Telemetry"
        ],
        "summary": "Get Telemetry History",
        "description": "Retrieves historical ranges of raw telemetry readings for all active sensors on a tank.",
        "operationId": "get_telemetry_history_api_v1_telemetry_history_get",
        "parameters": [
          {
            "name": "tank_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Tank Id"
            }
          },
          {
            "name": "start_time",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time",
              "title": "Start Time"
            }
          },
          {
            "name": "end_time",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time",
              "title": "End Time"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponse_TelemetryHistoryResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/users/{user_id}": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Get User Profile",
        "description": "Retrieves the registration properties for a single user.",
        "operationId": "get_user_profile_api_v1_users__user_id__get",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponse_UserResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/users/{user_id}/audit-logs": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Get User Audit Logs",
        "description": "Retrieves security and operations edit audit trails logged for the user.",
        "operationId": "get_user_audit_logs_api_v1_users__user_id__audit_logs_get",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "User Id"
            }
          },
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponse_List_AuditLogResponse__"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AuditLogResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "time": {
            "type": "string",
            "format": "date-time",
            "title": "Time"
          },
          "action": {
            "type": "string",
            "maxLength": 100,
            "title": "Action"
          },
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id"
          },
          "ip_address": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 45
              },
              {
                "type": "null"
              }
            ],
            "title": "Ip Address"
          },
          "user_agent": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Agent"
          },
          "details": {
            "additionalProperties": true,
            "type": "object",
            "title": "Details"
          }
        },
        "type": "object",
        "required": [
          "id",
          "time",
          "action",
          "user_id",
          "details"
        ],
        "title": "AuditLogResponse"
      },
...
      "BaseResponse_List_AuditLogResponse__": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          },
          "data": {
            "items": {
              "$ref": "#/components/schemas/AuditLogResponse"
            },
            "type": "array",
            "title": "Data"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "BaseResponse[List[AuditLogResponse]]"
      },
...
      "BaseResponse_TelemetryHistoryResponse_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          },
          "data": {
            "$ref": "#/components/schemas/TelemetryHistoryResponse"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "BaseResponse[TelemetryHistoryResponse]"
      },
      "BaseResponse_TelemetryLatestResponse_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          },
          "data": {
            "$ref": "#/components/schemas/TelemetryLatestResponse"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "BaseResponse[TelemetryLatestResponse]"
      },
...
      "BaseResponse_UserResponse_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          },
          "data": {
            "$ref": "#/components/schemas/UserResponse"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "BaseResponse[UserResponse]"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "TelemetryHistoryPoint": {
        "properties": {
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          },
          "value": {
            "type": "number",
            "title": "Value"
          },
          "sensor_id": {
            "type": "string",
            "format": "uuid",
            "title": "Sensor Id"
          }
        },
        "type": "object",
        "required": [
          "timestamp",
          "value",
          "sensor_id"
        ],
        "title": "TelemetryHistoryPoint"
      },
      "TelemetryHistoryResponse": {
        "properties": {
          "tankId": {
            "type": "string",
            "format": "uuid",
            "title": "Tankid"
          },
          "history": {
            "items": {
              "$ref": "#/components/schemas/TelemetryHistoryPoint"
            },
            "type": "array",
            "title": "History"
          }
        },
        "type": "object",
        "required": [
          "tankId",
          "history"
        ],
        "title": "TelemetryHistoryResponse"
      },
      "TelemetryLatestResponse": {
        "properties": {
          "tankId": {
            "type": "string",
            "format": "uuid",
            "title": "Tankid"
          },
          "capturedAt": {
            "type": "string",
            "format": "date-time",
            "title": "Capturedat"
          },
          "metrics": {
            "$ref": "#/components/schemas/TelemetryMetrics"
          }
        },
        "type": "object",
        "required": [
          "tankId",
          "capturedAt",
          "metrics"
        ],
        "title": "TelemetryLatestResponse"
      },
      "TelemetryMetrics": {
        "properties": {
          "temperature": {
            "type": "number",
            "title": "Temperature"
          },
          "ph": {
            "type": "number",
            "title": "Ph"
          },
          "dissolvedOxygen": {
            "type": "number",
            "title": "Dissolvedoxygen"
          },
          "salinity": {
            "type": "number",
            "title": "Salinity"
          },
          "ammonia": {
            "type": "number",
            "title": "Ammonia"
          },
          "turbidity": {
            "type": "number",
            "title": "Turbidity"
          }
        },
        "type": "object",
        "required": [
          "temperature",
          "ph",
          "dissolvedOxygen",
          "salinity",
          "ammonia",
          "turbidity"
        ],
        "title": "TelemetryMetrics"
      },
...
      "UserResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "email": {
            "type": "string",
            "maxLength": 150,
            "title": "Email"
          },
          "first_name": {
            "type": "string",
            "maxLength": 100,
            "title": "First Name"
          },
          "last_name": {
            "type": "string",
            "maxLength": 100,
            "title": "Last Name"
          },
          "role_id": {
            "type": "string",
            "format": "uuid",
            "title": "Role Id"
          }
        },
        "type": "object",
        "required": [
          "id",
          "email",
          "first_name",
          "last_name",
          "role_id"
        ],
        "title": "UserResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}